requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
matplotlib>=3.8
numpy>=1.26